+++ src/config/studio.php
@@ -48,5 +48,6 @@ return [
         //'episodeTxtField' => 'episodeTxt',
         //'podcastValueField' => 'podcastValue',
         //'episodeValueField' => 'episodeValue',
+        //'podrollField' => 'podroll',
     ],
 ];